feat: :fire: Remove Result screen, all uploads handled on questions page
