+++ App.tsx
@@ -5,7 +5,6 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import HomeScreen from "./screens/HomeScreen";
 import QuestionDisplay from "./screens/QuestionsScreen/questionPage";
-import Results from "./screens/ResultsScreen";
 //import dataTest from "./backend/dataTest";
 
 const Stack = createStackNavigator();
@@ -22,7 +21,6 @@ export default function App() {
       <Stack.Navigator>
         <Stack.Screen name="Home" component={HomeScreen} />
         <Stack.Screen name="Questions" component={QuestionDisplay} />
-        <Stack.Screen name="Results" component={Results} />
       </Stack.Navigator>
     </NavigationContainer>
   );
